Fixed inconsistent interface declarations, cleanup
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@3574 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/solvers/prop/prop.h
@@ -60,7 +60,7 @@ public:
   // variables
   virtual literalt new_variable()=0;
   virtual void set_variable_name(literalt a, const std::string &name) { }
-  virtual unsigned no_variables() const=0;
+  virtual size_t no_variables() const=0;
   bvt new_variables(unsigned width);
   
   // solving
